src/server/state/ create* => src/server/state/factory
